requirements.txt
streamlit
pandas
numpy
yfinance
sentence-transformers
transformers
faiss-cpu
scikit-learn
matplotlib
python-dateutil
requests
tqdm
python-dotenv
textblob
transformers
torch
plotly
fastapi
uvicorn
dash
dash-bootstrap-components
fastapi
uvicorn[standard]
requests
python-dotenv
yfinance
pandas
textblob
rich
google-genai
